Move image/document attachments code to a concern
This way we reduce some of the duplication in these classes.
